+++ changelog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Known Issues]
 - None. Please feel free to submit an issue via [GitHub](https://github.com/ryanblenis/MeshCentral-ScriptTask) if you find anything.
 
+## [0.0.6] - 2019-12-23
+### Added
+- Status return of a script as success/failure if the script does not produce output
+- Endpoint debugging toggle via console (plugin scripttask debug)
+### Fixed
+- Made some endpoint calls a little safer with try/catch to stop errors from killing the process
+
 ## [0.0.5] - 2019-12-22
 ### Added
 - NeDB support